+++ packages/minigame-std/CHANGELOG.md
@@ -532,6 +532,7 @@
   - `socket` - WebSocket
   - `storage` - 本地存储
 
+[2.5.0]: https://github.com/JiangJie/minigame-std/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/JiangJie/minigame-std/compare/v2.3.1...v2.4.0
 [2.3.1]: https://github.com/JiangJie/minigame-std/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/JiangJie/minigame-std/compare/v2.2.0...v2.3.0